Move cmd line handler
